Added textfield component to textfield package
